Changed ImageLoading from coil to database entry in homeScreen

+++ Kyoku/app/src/main/java/com/poulastaa/kyoku/data/model/database/table/ArtistPreviewSongRelation.kt
@@ -25,7 +25,7 @@ import androidx.room.ForeignKey
 )
 data class ArtistPreviewSongRelation(
     @ColumnInfo(index = true)
-    val artistId: Int,
+    val artistId: Long,
     @ColumnInfo(index = true)
     val songId: Long
 )
